keylogger.py
This updated version of the code includes the `key.ignore()` method call in the `keylogger` function to consume keyboard events and prevent them from being displayed on the screen. It also includes the disclaimer and user acceptance prompts, as well as the code to log keystrokes for a specified duration and save them to a log file.